書籍の削除時にダイアログによる確認がある#10

+++ intern-book-app/src/app/pages/book-list/book-list.component.ts
@@ -5,6 +5,8 @@ import { MatButtonModule } from '@angular/material/button';
 import { Book } from '../../types/book';
 import { BookCardComponent } from '../../components/book-card/book-card.component';
 import { BookService } from '../../book.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   standalone: true,
@@ -24,16 +26,28 @@ handleDelete($event: number) {
 }
   books: Book[] = []; // 書籍データを格納する配列
 
-  // BookService を注入
-  constructor(private bookService: BookService) { }
+  // BookService, MatDialogを注入
+  constructor(
+    private bookService: BookService,
+    private dialog: MatDialog
+  ){ }
 
   ngOnInit(): void {
     this.bookService.books$.subscribe(books => this.books = books)
   }
 
   onDeleteBook(bookId: number): void {
-    // 指定したIDの書籍を削除
-    this.bookService.deleteBook(bookId);
-  }
+    // 1. ダイアログを開く
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '300px',
+    });
 
+    // 2. ダイアログの結果を受け取る
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        // 指定したIDの書籍を削除
+        this.bookService.deleteBook(bookId);
+      }
+    });
+  }
 }
